Reorganized into separate directories
Things were getting crowded.
I considered moving over the example images, but I thought better of it
because users are likely to run the scripts themselves.  The Python
versions produce image files with the same names as the examples, and
I don't want to overwrite the example images.  An alternate approach
would be to move everything over and change the output filenames for
the Python scripts.  This is probably the thing to do.

Another advantage of this approach is that it allows me to write
commentary for each of the scripts by putting a readme in the
separate directories.